requirements.txt
accelerate==0.33.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
banks==2.2.0
beautifulsoup4==4.13.5
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.18
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
einops==0.8.1
faiss-cpu==1.12.0
fastapi==0.116.1
filelock==3.19.1
filetype==1.2.0
frozenlist==1.7.0
fsspec==2025.9.0
greenlet==3.2.4
griffe==1.14.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.23.2
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
llama-cloud==0.1.35
llama-cloud-services==0.6.54
llama-index==0.13.5
llama-index-cli==0.5.0
llama-index-core==0.13.5
llama-index-embeddings-huggingface==0.6.0
llama-index-embeddings-openai==0.5.0
llama-index-indices-managed-llama-cloud==0.9.3
llama-index-instrumentation==0.4.0
llama-index-llms-openai==0.5.4
llama-index-readers-file==0.5.3
llama-index-readers-llama-parse==0.5.0
llama-index-vector-stores-faiss==0.5.0
llama-index-workflows==1.3.0
llama-parse==0.6.54
llama_cpp_python==0.3.16
loguru==0.7.3
MarkupSafe==3.0.2
marshmallow==3.26.1
minijinja==2.12.0
mpmath==1.3.0
multidict==6.6.4
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
numpy==1.26.4
openai==1.106.1
packaging==25.0
pandas==2.2.3
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
propcache==0.3.2
psutil==7.0.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pypdf==6.0.0
pytest==8.4.1
pytest-asyncio==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.3
safetensors==0.6.2
scikit-learn==1.4.2
scipy==1.11.4
sentence-transformers==3.0.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
starlette==0.47.3
striprtf==0.0.26
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.19.1
torch==2.3.1
tqdm==4.67.1
transformers==4.43.4
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
wrapt==1.17.3
yarl==1.20.1